requirements.txt
# Local AI Companion - Dependencies

# Async HTTP client (for communicating with Ollama)
httpx>=0.25.0

# YAML configuration management
pyyaml>=6.0

# Text-to-Speech
edge-tts>=6.1.0      # Edge TTS (Microsoft cloud, fallback)
kokoro>=0.9.4        # Kokoro TTS (local, high quality)
f5-tts>=0.2.0        # F5-TTS voice cloning (Juri voice)
soundfile>=0.12.0    # For saving WAV files

# Speech-to-Text (ASR)
faster-whisper>=1.0.0  # Optimized Whisper (4x faster)
sounddevice>=0.4.0     # Microphone audio capture
huggingface_hub>=0.20.0  # For downloading optimized French models

# Web Server
fastapi>=0.104.0       # Async web framework
uvicorn>=0.24.0        # ASGI server
websockets>=12.0       # WebSocket support
python-multipart>=0.0.6  # For file uploads
torchaudio>=2.1.0      # Needed for XTTS/F5-TTS audio save

# Live2D Avatar Overlay
pywebview>=5.0.0       # Transparent window for Live2D avatar
pynput>=1.7.0          # Global hotkeys (F11, F12)
numpy>=1.24.0          # Audio analysis for lip-sync

# Desktop Companion (optional)
pystray>=0.19.0        # System tray icon
pillow>=10.0.0         # Image handling for tray icon

# Language Detection (for auto voice switching)
langdetect>=1.0.9      # Language detection
emoji>=2.0.0           # Emoji removal for TTS
